Extract non-selector behaviour into utils
