+++ client-v2/src/redux/modules/pageViewData/selectors.ts
@@ -1,19 +1,8 @@
 import { State } from '../../../types/State';
 import {
   PageViewDataPerFront,
-  PageViewDataPerCollection,
-  CollectionWithArticles,
-  ArticlePathAndId
+  PageViewDataPerCollection
 } from 'shared/types/PageViewData';
-import { ArticleFragment, Stages } from 'shared/types/Collection';
-import { DerivedArticle } from 'shared/types/Article';
-import {
-  selectArticleFragment,
-  selectSharedState,
-  createSelectArticlesInCollection,
-  createSelectArticleFromArticleFragment
-} from 'shared/selectors/shared';
-import { Store } from 'types/Store';
 
 const selectPageViewDataForArticlePath = (state: State, url: string) =>
   selectPageViewData(state);
@@ -40,9 +29,6 @@ const selectPageViewDataForCollection = (
   );
 };
 
-const selectAllArticleIds = createSelectArticlesInCollection();
-const selectArticle = createSelectArticleFromArticleFragment();
-
 const selectOpenCollectionsForFront = (
   allCollectionsInAFront: string[],
   openCollectionIds: string[]
@@ -52,54 +38,9 @@ const selectOpenCollectionsForFront = (
   );
 };
 
-// this is arguably not a selector so could be moved to a utils file?
-const selectCollectionsWithArticles = (
-  store: Store,
-  openCollectionIds: string[],
-  browsingStage: Stages = 'draft'
-): CollectionWithArticles[] => {
-  const state = store.getState();
-
-  return openCollectionIds.map((cId: string) => {
-    const allArticleIds: string[] = selectAllArticleIds(
-      selectSharedState(state),
-      {
-        collectionId: cId,
-        collectionSet: browsingStage,
-        includeSupportingArticles: false
-      }
-    );
-    const allArticleFrags: ArticleFragment[] = allArticleIds.map(articleId =>
-      selectArticleFragment(selectSharedState(state), articleId)
-    );
-    const allArticles: Array<DerivedArticle | undefined> = allArticleFrags.map(
-      frag => selectArticle(selectSharedState(state), frag.uuid)
-    );
-
-    const allArticlePathsAndIds: ArticlePathAndId[] = allArticles.reduce(
-      (acc, article) => {
-        if (article && article.urlPath) {
-          acc.push({
-            articlePath: article.urlPath,
-            articleId: article.uuid
-          });
-        }
-        return acc;
-      },
-      [] as ArticlePathAndId[]
-    );
-
-    return {
-      id: cId,
-      articles: allArticlePathsAndIds
-    };
-  });
-};
-
 export {
   selectPageViewData,
   selectPageViewDataForArticlePath,
   selectPageViewDataForCollection,
-  selectCollectionsWithArticles,
   selectOpenCollectionsForFront
 };
